backend/requirements.txt
# Django 核心
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5

# 数据库
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# 认证与安全
djangorestframework-simplejwt>=5.3.0
django-oauth-toolkit>=2.3.0
PyJWT>=2.8.0
cryptography>=41.0.0

# 缓存
django-redis>=5.4.0

# 异步任务
celery>=5.3.0
redis>=5.0.0

# API 文档
drf-spectacular>=0.27.0
drf-spectacular-sidecar>=2024.1.1

# 文件处理
Pillow>=10.2.0
openpyxl>=3.1.2
pandas>=2.1.4
xlrd>=2.0.1

# SSO 集成
requests>=2.31.0
wechatpy>=1.8.18

# 工具
python-dotenv>=1.0.0
gunicorn>=21.2.0

# 开发工具
django-debug-toolbar>=4.2.0
ipython>=8.18.0

# 日志
django-structlog>=6.0.0

# 导入导出
django-import-export>=3.3.0

# 树形结构支持
django-mptt>=0.15.0

# 定时任务
django-celery-beat>=2.5.0
django-celery-results>=2.5.1
